+++ domain/src/main/java/com/depromeet/team5/domain/review/ReviewService.java
@@ -1,8 +1,5 @@
-package com.depromeet.team5.service;
+package com.depromeet.team5.domain.review;
 
-import com.depromeet.team5.domain.review.ReviewUpdateValue;
-import com.depromeet.team5.domain.review.Review;
-import com.depromeet.team5.domain.review.ReviewCreateValue;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
