+++ domain/src/main/java/com/depromeet/team5/exception/BadRequestException.java
@@ -1,6 +1,6 @@
 package com.depromeet.team5.exception;
 
-public abstract class BadRequestException extends ApplicationException {
+public class BadRequestException extends ApplicationException {
     public BadRequestException(String message) {
         super(message);
     }
