+++ domain/src/main/java/com/depromeet/team5/service/impl/ReviewServiceImpl.java
@@ -8,9 +8,9 @@ import com.depromeet.team5.domain.store.Store;
 import com.depromeet.team5.domain.user.User;
 import com.depromeet.team5.exception.ReviewNotFoundException;
 import com.depromeet.team5.exception.StoreNotFoundException;
-import com.depromeet.team5.repository.ReviewRepository;
+import com.depromeet.team5.domain.review.ReviewRepository;
 import com.depromeet.team5.repository.StoreRepository;
-import com.depromeet.team5.service.ReviewService;
+import com.depromeet.team5.domain.review.ReviewService;
 import com.depromeet.team5.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
