+++ interfaces/src/test/java/com/depromeet/team5/integration/ReviewTest.java
@@ -5,9 +5,8 @@ import com.depromeet.team5.dto.*;
 import com.depromeet.team5.integration.api.ReviewTestController;
 import com.depromeet.team5.integration.api.StoreTestController;
 import com.depromeet.team5.integration.api.UserTestController;
-import com.depromeet.team5.repository.ReviewRepository;
+import com.depromeet.team5.domain.review.ReviewRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import jdk.nashorn.internal.ir.annotations.Ignore;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
